yolo/requirements.txt
confluent-kafka
Pillow
torch
opencv-python
ultralytics
bezier

gitpython
gitdb
smmap
setuptools==68.2.2
